+++ src/aioway/torch/nn/dense.py
@@ -2,7 +2,7 @@
 
 from torch import nn
 
-from .hop import NnLayerIter
+from .iters import NnLayerIter
 from .modules import NnInit, nn_init_dcls
 
 __all__ = ["Identity", "Linear", "Bilinear"]
